+++ server/util/images/deleteImages.ts
