+++ server/util/locations/deleteLocations.ts
